+++ src/main/java/be/baur/sdt/xpath/SDAXPath.java
@@ -10,6 +10,7 @@ import be.baur.sdt.xpath.function.CompareStringFunction;
 import be.baur.sdt.xpath.function.CurrentDateTimeFunction;
 import be.baur.sdt.xpath.function.DateTimeFunction;
 import be.baur.sdt.xpath.function.DocumentNodeFunction;
+import be.baur.sdt.xpath.function.FormatDateTimeFunction;
 import be.baur.sdt.xpath.function.LeftFunction;
 import be.baur.sdt.xpath.function.ParseSDAFunction;
 import be.baur.sdt.xpath.function.RenderSDAFunction;
@@ -45,6 +46,7 @@ public class SDAXPath extends BaseXPath {
 		ctx.registerFunction(SDT.W3CFUNCTIONS_NS_URI, "current-dateTime", new CurrentDateTimeFunction());
 		ctx.registerFunction(SDT.FUNCTIONS_NS_URI, "document-node", new DocumentNodeFunction());
 		ctx.registerFunction(SDT.FUNCTIONS_NS_URI, "dateTime", new DateTimeFunction());
+		ctx.registerFunction(SDT.FUNCTIONS_NS_URI, "format-dateTime", new FormatDateTimeFunction());
 		ctx.registerFunction(SDT.FUNCTIONS_NS_URI, "left", new LeftFunction());
 		ctx.registerFunction(SDT.FUNCTIONS_NS_URI, "parse-sda", new ParseSDAFunction());
 		ctx.registerFunction(SDT.FUNCTIONS_NS_URI, "right", new RightFunction());