+++ src/test/java/test/TestSDAXPath.java
@@ -193,12 +193,11 @@ public class TestSDAXPath {
 		t.so("S131", "sdt:dateTime('1968-02-28T12:00')", doc, "1968-02-28T12:00:00");
 		t.so("S132", "sdt:dateTime('1968-02-28T12:00+01:00')", doc, "1968-02-28T12:00:00+01:00");
 		t.so("S133", "sdt:dateTime('1968-02-28T12:00:00.000Z')", doc, "1968-02-28T12:00:00Z");
-		t.so("F134", "sdt:dateTime()", doc, "dateTime() requires exactly one argument.");
-		t.so("F135", "sdt:dateTime('abc')", doc, "dateTime() evaluation of 'abc' failed.");
-		t.so("S136", "fn:current-dateTime()", doc, CurrentDateTimeFunction.evaluate());
-		t.so("S137", "string(sdt:timestamp())", doc, String.format("%d", TimestampFunction.evaluate().longValue()));
-		t.so("F138", "sdt:timestamp(0)", doc, "timestamp() requires no arguments.");
-
+		t.so("F134", "sdt:dateTime('abc')", doc, "dateTime() evaluation of 'abc' failed.");
+		t.so("S135", "fn:current-dateTime()", doc, CurrentDateTimeFunction.evaluate());
+		t.so("S136", "string(sdt:timestamp())", doc, String.format("%d", TimestampFunction.evaluate().longValue()));
+		t.so("S137", "sdt:format-dateTime('1968-02-28T12:00','yyyy/MM/dd HH:mm')", doc, "1968/02/28 12:00");
+		t.so("S138", "sdt:format-dateTime(sdt:dateTime(0),'yyyyMMddHHmmss')", doc, "19700101000000");
 	}
 
 }